requirements.txt
PyGithub>=1.0,<=2.0
